backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = []

[dependency-groups]
dev = [
    "fastapi>=0.115.12",
    "google-search-results>=2.4.2",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "serpapi>=0.1.5",
    "uvicorn>=0.34.2",
]
